TINY-6487: Move blocking behaviour types into their own file
