+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/blocking/BlockingState.ts
@@ -1,13 +1,6 @@
 import { Cell, Singleton } from '@ephox/katamari';
-import { BehaviourState, nuState } from '../common/BehaviourState';
-
-export interface BlockingState extends BehaviourState {
-  // We don't actually store the blocker, just a callback to delete it
-  setBlocker: (destroy: () => void) => void;
-  clearBlocker: () => void;
-  getBlocked: () => boolean;
-  setBlocked: (blocked: boolean) => void;
-}
+import { nuState } from '../common/BehaviourState';
+import { BlockingState } from './BlockingTypes';
 
 export const init = (): BlockingState => {
   const blocker = Singleton.destroyable();
